Move Geometry subclasses into Geometry file
