--- ground/src/main/java/com/google/android/ground/model/geometry/LineString.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.android.ground.model.geometry
-
-import com.google.android.ground.util.toImmutableList
-import com.google.common.collect.ImmutableList
-import kotlinx.serialization.Serializable
-
-/** A sequence of two or more vertices modelling an OCG style line string. */
-@Serializable
-data class LineString(val coordinates: List<Coordinate>) : Geometry {
-  override val vertices: List<Point> = coordinates.map { Point(it) }.toImmutableList()
-}
